Allow gzip cytoBands
