--- packages/core/.gitignore
@@ -1,7 +1,0 @@
-.DS_Store
-node_modules
-coverage
-dist
-*.log
-*.swp
-
