+++ packages/core/data_adapters/CytobandAdapter/index.ts
@@ -1,0 +1,14 @@
+import { AdapterType } from '../../pluggableElementTypes'
+import PluginManager from '../../PluginManager'
+import configSchema from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'CytobandAdapter',
+        configSchema,
+        getAdapterClass: () => import('./CytobandAdapter').then(f => f.default),
+      }),
+  )
+}
