+++ packages/core/package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.17.9",
+    "@gmod/bgzf-filehandle": "^1.4.3",
     "@mui/icons-material": "^5.0.1",
     "@types/clone": "^2.0.0",
     "abortable-promise-cache": "^1.5.0",
